# src/main/java/co/com/cloud/computing/msbacklog/api/ProjectApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.34).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package co.com.cloud.computing.msbacklog.api;

import co.com.cloud.computing.msbacklog.model.Project;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.validation.Valid;
import java.util.List;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen",
        date = "2022-05-23T04:04:15.496Z[GMT]")
@Validated
public interface ProjectApi {

    @Operation(summary = "Retorna HTTP 200 al consultar exitosamente los proyectos.",
            description = "Ver todos proyecto.", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Proyectos Consultas Exitasamente",
                content = @Content(mediaType = "application/json", schema = @Schema(implementation = Project.class))),
        
        @ApiResponse(responseCode = "400", description = "Bad Request") })
    @RequestMapping(value = "/project",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Project>> projectGet();


    @Operation(summary = "Retorna HTTP 201 al crear exitosamente.", description = "Adicionar nuevo proyecto.", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "201", description = "Creado Exitasamente",
                content = @Content(mediaType = "application/json", schema = @Schema(implementation = Project.class))),
        
        @ApiResponse(responseCode = "400", description = "Bad Request") })
    @RequestMapping(value = "/project",
        produces = { "application/json" }, 
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Project> projectPost(@Parameter(in = ParameterIn.DEFAULT,
            description = "Cuerpo de la petición Project", required=true, schema=@Schema())
                                        @Valid @RequestBody Project body);

}